shamelessly steal the implementation of WatchStream
I can't use `WatchStream` directly, because it requires `T: Clone`,
and anyway I want to transform a stream of `Summary` into a stream
of `Result<Summary, !>`, because I'm required to return that result
because of the interface `tonic` generated for me.

On the other hand, `WatchStream` clearly works, so why reinvent the wheel?
Its source is easy to find: https://docs.rs/tokio-stream/latest/src/tokio_stream/wrappers/watch.rs.html#71-88

More-or-less copied that, making modifications as required for this
use case, and voila: the stream works!